fix: move error folder in sdk
