+++ packages/nc-gui/error/computed-type-paste.error.ts
@@ -1,7 +1,1 @@
-import { TypeConversionError } from './type-conversion.error'
-
-export class ComputedTypePasteError extends TypeConversionError {
-  constructor() {
-    super('Paste operation is not supported on the active cell')
-  }
-}
+export { ComputedTypePasteError } from 'nocodb-sdk'
