+++ packages/nc-gui/error/silent-type-conversion.error.ts
@@ -1,11 +1,1 @@
-import type { SuppressedError } from './suppressed.error'
-import { TypeConversionError } from './type-conversion.error'
-
-export class SilentTypeConversionError extends TypeConversionError implements SuppressedError {
-  constructor() {
-    super('')
-    this.isErrorSuppressed = true
-  }
-
-  isErrorSuppressed: boolean
-}
+export { SilentTypeConversionError } from 'nocodb-sdk'
