+++ packages/nc-gui/error/type-conversion.error.ts
@@ -1,5 +1,1 @@
-export class TypeConversionError extends TypeError {
-  constructor(message: string) {
-    super(message)
-  }
-}
+export { TypeConversionError } from 'nocodb-sdk'
