+++ packages/nocodb-sdk/src/lib/error/computed-type-paste.error.ts
@@ -1,0 +1,7 @@
+import { TypeConversionError } from './type-conversion.error'
+
+export class ComputedTypePasteError extends TypeConversionError {
+  constructor() {
+    super('Paste operation is not supported on the active cell')
+  }
+}
