+++ packages/nocodb-sdk/src/lib/error/suppressed.error.ts
@@ -1,0 +1,3 @@
+export interface SuppressedError {
+  isErrorSuppressed: boolean
+}
