+++ packages/nocodb-sdk/src/lib/error/type-conversion.error.ts
@@ -1,0 +1,5 @@
+export class TypeConversionError extends TypeError {
+  constructor(message: string) {
+    super(message)
+  }
+}
