+++ packages/nocodb-sdk/src/lib/index.ts
@@ -52,3 +52,4 @@ export * from '~/lib/is';
 export * from '~/lib/durationUtils';
 export * from '~/lib/sync';
 export * from '~/lib/columnHelper';
+export * from '~/lib/error';
